refactor to multiple plugins
