+++ src/plugins/unpkg-path-plugin.ts
@@ -1,12 +1,7 @@
 import * as esbuild from "esbuild-wasm";
-import axios from "axios";
-import localforage from "localforage";
 
-const fileCache = localforage.createInstance({
-  name: "fileCache",
-});
 
-export const unpkgPathPlugin = (inputCode: string) => {
+export const unpkgPathPlugin = () => {
   return {
     name: "unpkg-path-plugin",
     setup(build: esbuild.PluginBuild) {
@@ -32,35 +27,7 @@ export const unpkgPathPlugin = (inputCode: string) => {
         };
       });
 
-      build.onLoad({ filter: /.*/ }, async (args: any) => {
-        console.log("onLoad", args);
-
-        if (args.path === "index.js") {
-          return {
-            loader: "jsx",
-            contents: inputCode,
-          };
-        }
-
-        const cachedResult = await fileCache.getItem<esbuild.OnLoadResult>(
-          args.path
-        );
-
-        if (cachedResult) {
-          return cachedResult;
-        }
-
-        const { data, request } = await axios.get(args.path);
-
-        const result: esbuild.OnLoadResult = {
-          loader: "jsx",
-          contents: data,
-          resolveDir: new URL("./", request.responseURL).pathname,
-        };
-        await fileCache.setItem(args.path, result);
-
-        return result;
-      });
+      
     },
   };
 };
